pyproject.toml
[project]
name = "fastapi-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
]
